Move GraphQL to bfDb Subfolder for Improved Project Structure
## SUMMARY
This commit restructures the project by moving the GraphQL schema and related files from a standalone `apps/graphql` directory into a subfolder within `apps/bfDb`. The changes include updates to import paths across various files and the deletion of the old standalone GraphQL directory. This reorganization aims to consolidate GraphQL-related functionalities with database models and utilities, enhancing maintainability and coherence of the project structure.

Key changes:
- The directory `apps/graphql` has been removed, and its contents have been moved to `apps/bfDb/graphql`.
- Updated import paths in multiple files to reflect the new location of GraphQL files.
- Updated documentation and configuration files to reference the new location of GraphQL components.

## TEST PLAN
1. Ensure that all import paths related to GraphQL are updated correctly across the project.
2. Verify that the GraphQL server can be started without errors using the new file path.
3. Run existing test suites to ensure no functionality is broken due to the restructuring.
4. Manually test key GraphQL queries and mutations to confirm they operate as expected.
