+++ apps/bfDb/graphql/types/__TEMPLATE__.ts
@@ -1,7 +1,7 @@
 // import { objectType } from "nexus";
 // import { getLogger } from "packages/logger/logger.ts";
-// // import { graphqlNode } from "apps/graphql/types/graphqlBfNode.ts";
-// // import { graphqlBfNode } from "apps/graphql/types/graphqlBfNode.ts";
+// // import { graphqlNode } from "apps/bfDb/graphql/types/graphqlBfNode.ts";
+// // import { graphqlBfNode } from "apps/bfDb/graphql/types/graphqlBfNode.ts";
 // const _logger = getLogger(import.meta);
 
 // export const graphql$1Type = objectType({
